Added biocViews to help devtools/remotes
Currently "bluster" is listed as a depedency but is a Bioconductor package. Installations using `install_github()` will fail because it will not pull in Bioconductor dependencies. Adding biocViews hints devtools and remotes to look at Bioconductor and allows this package to be directly installed off github without manually installing bluster first.